+++ sdk/cosmos/azure-cosmos-kafka-connect/src/main/java/module-info.java
@@ -6,10 +6,10 @@ module com.azure.cosmos.kafka.connect {
 
     requires transitive com.azure.cosmos;
     requires kafka.clients;
-
+    requires connect.api;
+    requires com.fasterxml.jackson.module.afterburner;
 
     // public API surface area
     exports com.azure.cosmos.kafka.connect;
-
     uses com.azure.core.util.tracing.Tracer;
 }