+++ sdk/cosmos/azure-cosmos-kafka-connect/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.0.0-beta.1 (Unreleased)
 
 #### Features Added
+* Added Source connector. See [PR 39410](https://github.com/Azure/azure-sdk-for-java/pull/39410) 
 
 #### Breaking Changes